QR Code Scanning (Only Independent to AR)
QR code scanning now works but only independently of using AR features.
A library has been used. In order to get QR code scanning to occur
simultanouesly with AR scanning, there must be a common camera interface
between the two features. Currently, the AR library works using one
camera setup, but the QR library works with a completely different
camera setup. If this QR library is to be used in the future, scripts
must be modified to access the camera in the same way that is required
by ARToolkit.

Known Problems: Still takes a long time for Android camera to
initialise.